microblog-04-database/requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
pymongo
python-testdata
ForgeryPy3
